+++ CHANGELOG.md
@@ -34,6 +34,7 @@ This release depends on, and has been tested with, the following Mapbox dependen
 - Mapbox Android Core `v5.0.2` ([release notes](https://github.com/mapbox/mapbox-events-android/releases/tag/core-5.0.2))
 -core-5.0.2))
 
+- Fixed an issue with `NavigationView` that prevented camera from going into following state after starting a trip. [#6442](https://github.com/mapbox/mapbox-navigation-android/pull/6442)
 
 ## Mapbox Navigation SDK 2.9.0-alpha.4 - 30 September, 2022
 ### Changelog